+++ src/os_driver_nodelet.cpp
@@ -136,8 +136,8 @@ class OusterDriver : public OusterSensor {
                 throw std::runtime_error("min_range equal to or exceeds max_range!");
             }
             // convert to millimeters
-            uint64_t min_range = impl::ulround(min_range_m * 1000);
-            uint64_t max_range = impl::ulround(max_range_m * 1000);
+            uint32_t min_range = impl::ulround(min_range_m * 1000);
+            uint32_t max_range = impl::ulround(max_range_m * 1000);
             auto rows_step = pnh.param("rows_step", 1);
             processors.push_back(
                 PointCloudProcessorFactory::create_point_cloud_processor(point_type, info,
@@ -145,7 +145,7 @@ class OusterDriver : public OusterSensor {
                     organized, destagger, min_range, max_range, rows_step,
                     [this](PointCloudProcessor_OutputType msgs) {
                         for (size_t i = 0; i < msgs.size(); ++i)
-                        lidar_pubs[i].publish(*msgs[i]);
+                            lidar_pubs[i].publish(*msgs[i]);
                     }
                 )
             );
